requirements.txt
python-telegram-bot>=20.0
pyyaml>=6.0
requests>=2.32
